Backend/requirements.txt
asgiref==3.8.1
cffi>=1.16.0
cryptography==43.0.1
Django==5.0.7
django-cors-headers==4.4.0
django-environ==0.11.2
djangorestframework>=3.15.2
ez_setup==0.9
mysqlclient==2.2.4
pillow==10.3.0
pycparser==2.22
PyJWT>=2.8.0
sqlparse==0.5.0
django-ratelimit>=4.1.0
drf-spectacular>=0.26.5
psutil>=5.9.0
redis>=4.5.0
django-redis>=5.2.0
celery>=5.3.0
django-celery-beat>=2.5.0
django-celery-results>=2.5.0
